Fixed the pivot problem

+++ type.h
@@ -7,7 +7,8 @@ typedef struct ELEMENTS
   /* struct for each element in the matrix */
   int row;
   int col;
-  double datum;
+  
+  long double datum;
 
   
 }ELEMENT;
